pyproject.toml
[tool.poetry]
name = "test-creation-case"
version = "0.1.0"
description = ""
authors = ["Vitalii Magomedov <vitalikmag28@gmail.com>"]
readme = "README.md"
packages = [{include = "test_creation_case"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
